requirements.txt
faster-whisper==0.10.0
moviepy==1.0.3
pyannote.audio==3.1.1
python-dotenv==1.0.1
torch>=2.2.0
tqdm==4.66.2
numpy>=1.24.0
decorator>=4.4.2
imageio>=2.31.6
imageio-ffmpeg>=0.4.9
proglog>=0.1.10
